+++ data/db.go
@@ -19,19 +19,19 @@ func getConn() (bolt.Conn, error) {
 }
 
 // Player, Mob, Object, Room, Quest, ItemInventory
-func nextId(dataType string) int64{
+func nextId(dataType string) int{
 	conn, _ := getConn()
 	defer conn.Close()
 	data, _, _, _ := conn.QueryNeoAll(fmt.Sprintf("MATCH (r:%[1]s) RETURN COALESCE(MAX(r.%[2]s_id), 0)", strings.ToLower(dataType), dataType), nil)
-	return data[0][0].(int64)+1
+	return data[0][0].(int)+1
 }
 
 // Player, Mob, Object, Room, Quest, ItemInventory
-func nextLinkId(dataType string) int64{
+func nextLinkId(dataType string) int{
 	conn, _ := getConn()
 	defer conn.Close()
 	data, _, _, _ := conn.QueryNeoAll(fmt.Sprintf("MATCH ()-[r:%[1]s]->() RETURN COALESCE(MAX(r.%[2]s_id), 0)", strings.ToLower(dataType), dataType), nil)
-	return data[0][0].(int64)+1
+	return data[0][0].(int)+1
 }
 
 //
@@ -114,7 +114,7 @@ func consumeRows(rows bolt.Rows, st bolt.Stmt) {
 	_, _, err = rows.NextNeo()
 	handleError(err)
 	fmt.Printf("COLUMNS: %#v\n", rows.Metadata()["fields"].([]interface{})) // COLUMNS: n.foo,n.bar
-	fmt.Printf("FIELDS: %d %f\n", data[0].(int64), data[1].(float64))       // FIELDS: 1 2.2
+	fmt.Printf("FIELDS: %d %f\n", data[0].(int), data[1].(float64))       // FIELDS: 1 2.2
 
 	st.Close()
 }
